// grpc/tag_pb.d.ts
export class Tag extends jspb.Message {
  getTagId(): number;
  setTagId(value: number): Tag;

  getTagName(): string;
  setTagName(value: string): Tag;

  getStatus(): number;
  setStatus(value: number): Tag;

  getCreateuserId(): number;
  setCreateuserId(value: number): Tag;

  getUpdateuserId(): number;
  setUpdateuserId(value: number): Tag;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): Tag.AsObject;
  static toObject(includeInstance: boolean, msg: Tag): Tag.AsObject;
  static serializeBinaryToWriter(message: Tag, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): Tag;
  static deserializeBinaryFromReader(message: Tag, reader: jspb.BinaryReader): Tag;
}

export namespace Tag {
  export type AsObject = {
    tagId: number,
    tagName: string,
    status: number,
    createuserId: number,
    updateuserId: number,
  }
}
// ...
export class CreateTagRequest extends jspb.Message {
  getTag(): Tag | undefined;
  setTag(value?: Tag): CreateTagRequest;
  hasTag(): boolean;
  clearTag(): CreateTagRequest;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): CreateTagRequest.AsObject;
  static toObject(includeInstance: boolean, msg: CreateTagRequest): CreateTagRequest.AsObject;
  static serializeBinaryToWriter(message: CreateTagRequest, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): CreateTagRequest;
  static deserializeBinaryFromReader(message: CreateTagRequest, reader: jspb.BinaryReader): CreateTagRequest;
}

export namespace CreateTagRequest {
  export type AsObject = {
    tag?: Tag.AsObject,
  }
}
// ...
export class ListTagRequest extends jspb.Message {
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): ListTagRequest.AsObject;
  static toObject(includeInstance: boolean, msg: ListTagRequest): ListTagRequest.AsObject;
  static serializeBinaryToWriter(message: ListTagRequest, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): ListTagRequest;
  static deserializeBinaryFromReader(message: ListTagRequest, reader: jspb.BinaryReader): ListTagRequest;
}

export namespace ListTagRequest {
  export type AsObject = {
  }
}